please dont

+++ hackTues/Assets/Scripts/AbstractFlower.cs
@@ -20,7 +20,7 @@ public class AbstractFlower: MonoBehaviour
 
     public float soil_req;
     public float growTime;
-    public float life;
+    public int life;
 
     public float flowerCost;
